+++ sdk/spring-boot-starter-test/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-kalix.springboot.KalixConfigurationTest
+kalix.spring.boot.KalixConfigurationTest
