+++ sdk/spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-kalix.springboot.KalixConfiguration
+kalix.spring.boot.KalixConfiguration
