+++ sdk/spring-sdk/src/it/java/com/example/Main.java
@@ -16,7 +16,7 @@
 
 package com.example;
 
-import kalix.springboot.KalixConfiguration;
+import kalix.spring.boot.KalixConfiguration;
 import kalix.springsdk.annotations.Acl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
